feat: reorganize block action schemas
