+++ backend/src/main/java/ch/nova_omnia/lernello/dto/request/block/create/AIGeneratedQuestionBlockRequest.java
@@ -5,7 +5,7 @@ import java.util.UUID;
 import jakarta.validation.constraints.NotNull;
 
 public record AIGeneratedQuestionBlockRequest(
-                                              @NotNull UUID theoryBlockIUuid,
-                                              @NotNull UUID questionBlockUuid
+                                              @NotNull UUID theoryBlockId,
+                                              @NotNull UUID questionBlockId
 ) {
 }
